fix: Change DIV operator to return Integer instead of Numeric
The DIV (integer division) operator was incorrectly returning Numeric type,
which caused results to display as decimals (e.g., 0.000, 1.000) instead of
integers (e.g., 0, 1).

Changes:
- Line 58: Fast path for Integer DIV Integer now returns Integer
- Line 77: ExactNumeric DIV result returns Integer
- Lines 78-79: ApproximateNumeric and Numeric DIV results return Integer

This fix resolves 99 test failures (49.5% of all failures) across:
- random/aggregates
- random/expr
- random/select
- index/random

Closes #1723

+++ crates/vibesql-executor/src/evaluator/operators/arithmetic/division.rs
@@ -55,7 +55,7 @@ impl Division {
             if *b == 0 {
                 return Ok(SqlValue::Null);
             }
-            return Ok(Numeric((a / b) as f64));
+            return Ok(Integer(a / b));
         }
 
         // Use helper for type coercion
@@ -74,9 +74,9 @@ impl Division {
 
         // Integer division truncates toward zero
         match coerced {
-            super::CoercedValues::ExactNumeric(a, b) => Ok(Numeric((a / b) as f64)),
-            super::CoercedValues::ApproximateNumeric(a, b) => Ok(Numeric((a / b).trunc())),
-            super::CoercedValues::Numeric(a, b) => Ok(Numeric((a / b).trunc())),
+            super::CoercedValues::ExactNumeric(a, b) => Ok(Integer(a / b)),
+            super::CoercedValues::ApproximateNumeric(a, b) => Ok(Integer((a / b).trunc() as i64)),
+            super::CoercedValues::Numeric(a, b) => Ok(Integer((a / b).trunc() as i64)),
         }
     }
 }
